Related to #423

Some performance improvements.

Benchmark results:
```
differenceInMilliseconds OLD x 8,603,205 ops/sec ±0.45% (98 runs sampled)
differenceInMilliseconds NEW x 48,500,068 ops/sec ±0.17% (99 runs sampled)

addHours OLD x 7,637,610 ops/sec ±0.20% (100 runs sampled)
addHours NEW x 14,971,984 ops/sec ±0.22% (97 runs sampled)

getTime OLD x 12,509,457 ops/sec ±0.31% (97 runs sampled)
getTime NEW x 60,525,215 ops/sec ±0.17% (99 runs sampled)

toDate OLD x 12,572,094 ops/sec ±0.18% (96 runs sampled)
toDate NEW x 14,607,214 ops/sec ±0.36% (98 runs sampled)

// related functions
differenceInMinutes OLD x 7,251,437 ops/sec ±0.24% (95 runs sampled)
differenceInMinutes NEW x 42,027,626 ops/sec ±2.13% (92 runs sampled)

differenceInHours OLD x 7,286,628 ops/sec ±0.18% (96 runs sampled)
differenceInHours NEW x 45,942,881 ops/sec ±0.32% (97 runs sampled)

addHours OLD x 7,662,544 ops/sec ±0.23% (98 runs sampled)
addHours NEW x 14,822,585 ops/sec ±0.67% (98 runs sampled)

addWeeks OLD x 6,617,463 ops/sec ±0.17% (100 runs sampled)
addWeeks NEW x 7,185,398 ops/sec ±0.25% (100 runs sampled)

getUnixTime OLD x 12,313,013 ops/sec ±0.56% (98 runs sampled)
getUnixTime NEW x 44,588,446 ops/sec ±0.95% (93 runs sampled)
```

Benchmark code:
```js
import Benchmark from 'benchmark';
import fnsOld from 'date-fns';
import fnsNew from './date-fns-new/index.js';

const leftDate = new Date();
const rightDate = fnsOld.addHours(new Date(), 453);
const addHours = 14;

const testsFunctions = [
  // changed functions
  ['differenceInMilliseconds', [leftDate, rightDate]],
  ['addHours', [leftDate, addHours]],
  ['getTime', [leftDate]],
  ['toDate', [leftDate]],
  // some related functions
  ['differenceInMinutes', [leftDate, rightDate]],
  ['differenceInHours', [leftDate, rightDate]],
  ['addHours', [leftDate, addHours]],
  ['addWeeks', [leftDate, addHours]],
  ['getUnixTime', [leftDate]],
];

testsFunctions.forEach(([funcName, args]) => {
  const suite = new Benchmark.Suite;

  suite
    .add(`${funcName} OLD`, () => fnsOld[funcName](...args))
    .add(`${funcName} NEW`, () => fnsNew[funcName](...args))
    .on('cycle', (event) => console.log(String(event.target)))
    .on('complete', () => console.log(''))
    .run();
});
```